feat comment(update, delete) errors

+++ server/controller/posts.js
@@ -23,7 +23,7 @@ export async function updatePost(req, res) {
     return res.sendStatus(403);
   }
 
-  const updated = await postsRepository.update(text, id);
+  const updated = await postsRepository.update(text, postId);
   res.status(200).json(updated);
 }
 
@@ -54,6 +54,15 @@ export async function updateComment(req, res) {
   const { text } = req.body;
   const postId = req.params.postid;
   const commentId = req.query.commentId;
+  const found = await postsRepository.getComment(postId, commentId);
+
+  if (found[0].comments.length === 0) {
+    return res.Status(404).json({ message: 'Can not find this comment' });
+  }
+  if (found[0].comments[0].author.toString() !== req.userId) {
+    return res.sendStatus(403);
+  }
+
   const updated = await postsRepository.modifyComment(text, postId, commentId);
   res.status(200).json(updated);
 }
@@ -61,6 +70,16 @@ export async function updateComment(req, res) {
 export async function deleteComment(req, res) {
   const postId = req.params.postid;
   const commentId = req.query.commentId;
+
+  const found = await postsRepository.getComment(postId, commentId);
+
+  if (found[0].comments.length === 0) {
+    return res.Status(404).json({ message: 'Can not find this comment' });
+  }
+  if (found[0].comments[0].author.toString() !== req.userId) {
+    return res.sendStatus(403);
+  }
+
   await postsRepository.deleteComment(postId, commentId);
   res.sendStatus(204);
 }
